+++ backend/users/Touir.ExpensesManager.Users/Services/Contracts/IAuthenticationService.cs
@@ -13,6 +13,6 @@ namespace Touir.ExpensesManager.Users.Services.Contracts
         Task<bool> ChangePasswordAsync(string email, string oldPassword, string newPassword);
         Task<bool> ResetPasswordAsync(string email, string verificationHash, string newPassword);
 
-        Task<bool> RequestPasswordResetAsync(string email);
+        Task<bool> RequestPasswordResetAsync(string email, string appCode);
     }
 }